Done u/logic array example: todo use logic_vector
