+++ vhpidirect/cinterface/logicvector/run.sh
@@ -6,5 +6,6 @@ set -e
 
 ghdl -a -O0 -g tb.vhd
 ghdl -e -O0 -g -Wl,-ggdb3 -Wl,-I./ -Wl,main.c tb &&
-valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --verbose --log-file=valgrind-out.txt ./tb &&
-cat valgrind-out.txt | grep -A 4 "ERROR SUMMARY"
+./tb
+#valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --verbose --log-file=valgrind-out.txt ./tb &&
+#cat valgrind-out.txt | grep -A 4 "LEAK SUMMARY"
